Lesson#12 - Translate, security, voter, KnpMenu, BreadCrumbs

+++ src/Controller/UserController.php
@@ -9,9 +9,20 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class UserController extends AbstractController
 {
+    private $translator;
+
+    /**
+     * @param TranslatorInterface $translator
+     */
+    public function __construct(TranslatorInterface $translator)
+    {
+        $this->translator = $translator;
+    }
+
     /**
      * @Route("/user/list", name="user_list")
      * @return Response
@@ -28,7 +39,7 @@ class UserController extends AbstractController
     }
 
     /** @param int $id
-     * @Route("/user_articles/{id}/", name="user_articles", requirements={"id" = "\d+"}, defaults={"id" = 1})
+     * @Route("/user_articles/{id}/", name="user_articles", requirements={"id" : "\d+"}, defaults={"id" = 1})
      * @param ArticleRepository $articles
      * @return Response
      */
@@ -43,7 +54,7 @@ class UserController extends AbstractController
 
     /**
      * @param User $user
-     * @Route("/user/{id}", methods={"GET", "POST"}, name="user_show", requirements={"id" = "\d+"}, defaults={"id" = 1})
+     * @Route("/user/{id}", methods={"GET", "POST"}, name="user_show", requirements={"id" : "\d+"}, defaults={"id" = 1})
      * @return Response
      */
     public function userShow(User $user): Response
@@ -60,6 +71,8 @@ class UserController extends AbstractController
      */
     public function userNew(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $user = new User();
 
         $form = $this->createForm(UserType::class, $user);
@@ -81,11 +94,13 @@ class UserController extends AbstractController
     /**
      * @param Request $request
      * @param User $user
-     * @Route("/user/edit/{id}/", name="user_edit", requirements={"id" = "\d+"}, defaults={"id" = 1})
+     * @Route("/user/edit/{id}/", name="user_edit", requirements={"id" : "\d+"}, defaults={"id" = 1})
      * @return Response
      */
     public function userEdit(Request $request, User $user): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $form = $this->createForm(UserType::class, $user);
 
         $form->handleRequest($request);
@@ -104,11 +119,15 @@ class UserController extends AbstractController
 
     /**
      * @param User $user
-     * @Route("/user/delete/{id}", name="user_delete", requirements={"id" = "\d+"})
+     * @Route("/user/delete/{id}", name="user_delete", requirements={"id" : "\d+"})
      * @return Response
      */
     public function userDelete(User $user): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
+        $user->getComments()->clear();
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($user);
         $em->flush();
